+++ src/ioc.config.ts
@@ -3,8 +3,8 @@ import "reflect-metadata";
 import {container} from "tsyringe";
 import {CustomRulePathManager} from './lib/CustomRulePathManager';
 import {DefaultRuleManager} from './lib/DefaultRuleManager';
-import {JavascriptEslintEngine} from './lib/eslint/EslintEngine';
-import {TypescriptEslintEngine} from './lib/eslint/EslintEngine';
+import {JavascriptEslintEngine} from './lib/eslint/EslintEngineImpl';
+import {TypescriptEslintEngine} from './lib/eslint/EslintEngineImpl';
 import {PmdEngine} from './lib/pmd/PmdEngine';
 import {RuleManager} from './lib/RuleManager';
 import {RulePathManager} from './lib/RulePathManager';
